+++ my/tinder/android.py
@@ -11,8 +11,9 @@ from pathlib import Path
 import sqlite3
 from typing import Sequence, Iterator, Union, Dict, List, Mapping
 
-from my.core import Paths, get_files, Res, assert_never, stat, Stats, datetime_aware, make_logger
+from my.core import Paths, get_files, Res, stat, Stats, datetime_aware, make_logger
 from my.core.common import unique_everseen
+from my.core.compat import assert_never
 from my.core.error import echain
 from my.core.sqlite import sqlite_connection
 import my.config
