+++ setup.py
@@ -5,6 +5,7 @@ from setuptools import setup, find_namespace_packages # type: ignore
 
 INSTALL_REQUIRES = [
     'pytz',           # even though it's not needed by the core, it's so common anyway...
+    'typing-extensions',  # one of the most common pypi packages, ok to depend for core
     'appdirs',        # very common, and makes it portable
     'more-itertools', # it's just too useful and very common anyway
     'decorator'     , # less pain in writing correct decorators. very mature and stable, so worth keeping in core
